+++ R/minimum.spanning.tree.R
@@ -53,7 +53,7 @@ minimum.spanning.tree <- function(
 #'
 #' @param graph The graph object to analyze.
 #' @param weights Numeric vector giving the weights of the edges in the
-#'   graph. The order is determined by the edge ids. This is ignored if the
+#'   graph. The order is determined by the edge IDs. This is ignored if the
 #'   `unweighted` algorithm is chosen. Edge weights are interpreted as
 #'   distances.
 #' @param algorithm The algorithm to use for calculation. `unweighted` can